Modified Diag method
