+++ src/operator/numpy/np_matrix_op.cc
@@ -36,6 +36,7 @@ DMLC_REGISTER_PARAMETER(NumpyMoveaxisParam);
 DMLC_REGISTER_PARAMETER(NumpyRot90Param);
 DMLC_REGISTER_PARAMETER(NumpyReshapeParam);
 DMLC_REGISTER_PARAMETER(NumpyXReshapeParam);
+DMLC_REGISTER_PARAMETER(NumpyDiagParam);
 
 
 bool NumpyTransposeShape(const nnvm::NodeAttrs& attrs,
@@ -1301,6 +1302,28 @@ NNVM_REGISTER_OP(_npi_hsplit_backward)
     return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
 })
 .set_attr<FCompute>("FCompute<cpu>", HSplitOpBackward<cpu>);
+
+NNVM_REGISTER_OP(_np_diag)
+.set_attr_parser(ParamParser<NumpyDiagParam>)
+.set_num_inputs(1)
+.set_num_outputs(1)
+.set_attr<nnvm::FListInputNames>("FListInputNames",
+    [](const NodeAttrs &attrs) {
+    return std::vector<std::string>{"data"};
+})
+.set_attr<mxnet::FInferShape>("FInferShape", NumpyDiagOpShape)
+.set_attr<nnvm::FInferType>("FInferType", NumpyDiagOpType)
+.set_attr<FCompute>("FCompute<cpu>", NumpyDiagOpForward<cpu>)
+.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseNone{"_backward_diag"})
+.add_argument("data", "NDArray-or-Symbol", "Input ndarray")
+.add_arguments(NumpyDiagParam::__FIELDS__());
+
+NNVM_REGISTER_OP(_backward_np_diag)
+.set_attr_parser(ParamParser<NumpyDiagParam>)
+.set_num_inputs(1)
+.set_num_outputs(1)
+.set_attr<nnvm::TIsBackward>("TIsBackward", true)
+.set_attr<FCompute>("FCompute<cpu>", NumpyDiagOpBackward<cpu>);
 
 }  // namespace op
 }  // namespace mxnet
